コンテンツスクリプトの処理のコンポーネント化
