+++ src/content.ts
@@ -1,135 +1,8 @@
-import { randChar, pwLength, copiedMessage, copiedMessageShowMs } from './constants';
-import { Notyf } from 'notyf';
-import tippy from 'tippy.js';
+import { copyComputedCssText, showToast } from './util';
+import { PwGenerateButton } from './components/content/PwGenerateButton';
+import { PackUpWithPwButton } from './components/content/PackUpWithPwButton';
 import 'notyf/notyf.min.css';
 import 'tippy.js/dist/tippy.css';
-
-const generatePw = (): string => {
-  return Array.from(crypto.getRandomValues(new Uint32Array(pwLength))).map((n) => randChar[n % randChar.length]).join('');
-};
-
-const copyComputedCssText = (target: HTMLElement, ignoreProps: Array<string>): string => {
-  // スタイルの取得
-  const ignorePropName: RegExp | string = ignoreProps !== undefined
-    ? new RegExp(`^(${ignoreProps.join('|')})$`)
-    : '';
-  const targetStyle: CSSStyleDeclaration = getComputedStyle(target);
-  let computedCssText: string = '';
-
-  // 取得したスタイルから必要なスタイルのみを抽出しCSSTextを組み立て
-  for (let i = 0; i < targetStyle.length; i++) {
-    const propName: string = targetStyle.item(i);
-    const propVal: string = targetStyle.getPropertyValue(propName);
-
-    if ((ignorePropName !== '' && (ignorePropName as RegExp).test(propName)) || propVal === '') {
-      continue;
-    }
-
-    computedCssText += `${propName}: ${propVal};`;
-  }
-
-  return computedCssText;
-}
-
-const copyToClipboard = async (copyText: string, tooltipTargetElement: HTMLElement): Promise<void> => {
-  const isCopyToClipboard = (await chrome.storage.sync.get('isCopyToClipboard'))['isCopyToClipboard'];
-  if (!isCopyToClipboard) {
-    return;
-  }
-
-  try {
-    await navigator.clipboard.writeText(copyText);
-    showCopiedNotice(tooltipTargetElement);
-  }
-  catch (e) {
-    alert('クリップボードへのコピーに失敗しました: ' + e);
-  }
-};
-
-const showCopiedTooltip = (targetElement: HTMLElement): void => {
-  let hideTimerId: number | null = null;
-  const tippyInstance = tippy(targetElement, {
-    content: copiedMessage,
-    trigger: 'manual',
-    // デフォルトだとmargin-top分上に表示されるためoffsetでその分を下げる
-    offset: [0, -(Number(targetElement.style.marginTop))],
-    onShown() {
-      hideTimerId = window.setTimeout(() => { tippyInstance.hide(); }, copiedMessageShowMs);
-    },
-    onHidden() {
-      if (hideTimerId !== null) {
-        window.clearTimeout(hideTimerId);
-      }
-      tippyInstance.destroy();
-    },
-  });
-
-  tippyInstance.show();
-};
-
-type ToastType = 'Update' | 'Copied';
-const notyf = new Notyf({
-  // 共通設定
-  ripple: false,
-  position: {
-    x: 'center',
-    y: 'top',
-  },
-  dismissible: true,
-  types: [
-    {
-      type: 'info',
-      duration: 0, // 自動で閉じない
-      background: '#5888e0',
-      icon: {
-        className: 'pwgen-toast__icon',
-        tagName: 'div',
-      },
-      className: 'pwgen-toast',
-    },
-    {
-      type: 'success',
-      duration: copiedMessageShowMs,
-      className: 'pwgen-toast',
-    },
-  ],
-});
-
-const showToast = (toastType: ToastType): void => {
-  if (toastType === 'Update') {
-    notyf.open({
-      type: 'info',
-      message: `
-        <p>ギガファイル便DLパスジェネレーターがアップデートされました！</p>
-        <p>詳しくは<a href="https://github.com/woorld/gigafile-pwgen/releases/" target="_blank" rel="noopener">こちら</a>をご覧ください。</p>
-      `,
-    });
-    return;
-  }
-  notyf.open({
-    type: 'success',
-    message: copiedMessage,
-  });
-};
-
-const isUploadedFile = (uploadFileArea: Element): boolean => uploadFileArea.querySelector<HTMLInputElement>('.file_info_url.url')!.value !== '';
-
-const showCopiedNotice = async (tooltipTargetElement: HTMLElement): Promise<void> => {
-  const copiedNoticeType = (await chrome.storage.sync.get('copiedNoticeType'))['copiedNoticeType'];
-  switch (copiedNoticeType) {
-    // 'none'の場合は何もしない
-    case 'tooltip':
-      showCopiedTooltip(tooltipTargetElement);
-      return;
-    case 'toast':
-      showToast('Copied');
-      return;
-    case 'both':
-      showCopiedTooltip(tooltipTargetElement);
-      showToast('Copied');
-      return;
-  }
-};
 
 chrome.storage.sync.get(['isEnable'], (result) => {
   if (!result.isEnable) {
@@ -161,80 +34,14 @@ chrome.storage.sync.get(['isEnable'], (result) => {
   });
 
   // 「まとめる」ボタンの横に「パスワード付きでまとめる」ボタンを追加
-  const buttonPackUp: HTMLElement = document.getElementById('matomete_btn')!;
-  const buttonPackUpWithPw = document.createElement('button');
-
-  // 「まとめる」ボタンに合わせてspanでボタン内テキストを作成し追加
-  const buttonText: HTMLElement = document.createElement('span');
-  buttonText.textContent = 'パスワード付きでまとめる';
-
-  let buttonCssText: string = copyComputedCssText(buttonPackUp, ['width', 'inline-size', 'padding']) + 'padding: 5px;';
+  const packUpButton: HTMLElement = document.getElementById('matomete_btn')!;
+  let packUpButtonCssText: string = copyComputedCssText(packUpButton, ['width', 'inline-size', 'padding']) + 'padding: 5px;';
   if (navigator.userAgent.match(/Mobile/)) {
-    buttonCssText += 'width:100%;';
+    packUpButtonCssText += 'width:100%;';
   }
 
-  buttonPackUpWithPw.style.cssText = buttonCssText;
-  buttonPackUpWithPw.className = 'pwgen-packup-with-pw';
-  buttonPackUpWithPw.appendChild(buttonText);
-
-  buttonPackUpWithPw.addEventListener('click', async () => {
-    const files = document.getElementsByClassName('file_info')!;
-    let existsUploadedFile = false;
-    let existsUploadPausingFile = false;
-
-    if (files.length <= 0) {
-      alert('ファイルを選択してください。');
-      return;
-    }
-
-    for (const file of files) {
-      const buttonCancelStatus = file.getElementsByClassName('cancel')[0].getAttribute('value');
-
-      // アップロード完了済ファイル・中断中ファイルはスキップ
-      if (isUploadedFile(file)) {
-        existsUploadedFile = true;
-        continue;
-      }
-      if (buttonCancelStatus === 'on') {
-        existsUploadPausingFile = true;
-        continue;
-      }
-
-      alert('アップロードが完了していないファイルがあります。\n完了してから再度お試しください。');
-      return;
-    }
-
-    // アップロード完了済ファイルがない（中断中ファイルしかない）場合はエラーになるため処理を抜ける
-    if (!existsUploadedFile) {
-      alert('全ファイルのアップロードが中断されています。\nいずれかのファイルのアップロードを再開し、アップロード完了後に再度お試しください。');
-      return;
-    }
-
-    if (existsUploadPausingFile) {
-      alert('アップロード未完了かつ中断中のファイルはまとめられません。\n次に表示されるアラートでは全ファイルをまとめた旨が通知されますが、実際にはアップロード済みのファイル（アップロード完了後に中断したものを含む）のみがまとめられていますのでご注意ください。');
-    }
-
-    const pw: string = generatePw();
-    const obsConfig = {
-      attributes: true,
-    };
-    const obsPackUp = new MutationObserver(async (mutationsList) => {
-      const dlUrl = (mutationsList[0].target as HTMLInputElement).attributes[4].value;
-      const copyText = `${dlUrl}\nダウンロードパスワード：${pw}`;
-
-      await copyToClipboard(copyText, buttonPackUpWithPw);
-
-      obsPackUp.disconnect();
-    });
-
-    obsPackUp.observe(document.getElementById('matomete_url')!, obsConfig);
-
-    // パスを入力して設定ボタンを押下
-    (document.getElementById('zip_dlkey') as HTMLInputElement).value = pw;
-    buttonPackUp.click();
-  });
-
-  buttonPackUp.parentNode!.appendChild(buttonPackUpWithPw);
+  const packUpWithPwButton = PackUpWithPwButton(packUpButtonCssText, packUpButton);
+  packUpButton.parentNode!.appendChild(packUpWithPwButton);
 
   const target = document.getElementById('file_list')!;
   const obsConfig = {
@@ -249,29 +56,10 @@ chrome.storage.sync.get(['isEnable'], (result) => {
           return;
         }
 
-        const button = document.createElement('button');
+        const pwGenerateButtonCssText = copyComputedCssText(document.getElementsByClassName('set_dlkey')[0] as HTMLButtonElement, ['width', 'inline-size']);
+        const pwGenerateButton = PwGenerateButton(pwGenerateButtonCssText, uploadFileArea);
 
-        button.textContent = 'PW生成・設定';
-        button.style.cssText = copyComputedCssText(document.getElementsByClassName('set_dlkey')[0] as HTMLButtonElement, ['width', 'inline-size']);
-
-        button.addEventListener('click', async () => {
-          if (!isUploadedFile(uploadFileArea)) {
-            alert('ファイルのアップロードが完了していません。\n完了してから再度お試しください。');
-            return;
-          }
-
-          const pw = generatePw();
-          // パスを入力して設定ボタンを押下
-          (uploadFileArea.getElementsByClassName('dlkey_inp')[0] as HTMLInputElement).value = pw;
-          (uploadFileArea.getElementsByClassName('set_dlkey gfbtn')[0] as HTMLElement).click();
-
-          const dlUrl = (uploadFileArea.getElementsByClassName('file_info_url url')[0] as HTMLInputElement).value;
-          const copyText = `${dlUrl}\nダウンロードパスワード：${pw}`;
-
-          await copyToClipboard(copyText, button);
-        });
-
-        uploadFileArea.getElementsByClassName('dlkey')[0].appendChild(button);
+        uploadFileArea.getElementsByClassName('dlkey')[0].appendChild(pwGenerateButton);
       });
     };
   });
